KNOX-2511 - Consolidate HA Dispatches
